phonehub: Fix signed vs unsigned comparison warning
shell-encryption and private_membership leaked some configs
which disable some default warnings.

In order to re-enable all these warnings for chrome/browser/*
in this CL crrev.com/c/3303510, all warnings need to be fixed.

This CL is fixing the comparison between two variables to have
the same type (i.e. either both are signed or unsigned integers).

Bug: chromium:1212837, b:201763017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3358241
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#960347}

+++ chrome/browser/ash/phonehub/browser_tabs_metadata_fetcher_impl.cc
@@ -26,7 +26,7 @@ GetSortedMetadataWithoutFavicons(const sync_sessions::SyncedSession* session) {
     for (const std::unique_ptr<sessions::SessionTab>& tab : window.tabs) {
       int selected_index = tab->normalized_navigation_index();
 
-      if (selected_index + 1 > tab->navigations.size() ||
+      if (selected_index + 1 > static_cast<int>(tab->navigations.size()) ||
           tab->navigations.empty()) {
         continue;
       }
